+++ lib/scheduler.dart
@@ -8,6 +8,7 @@ library dart_web_toolkit_scheduler;
 
 import 'dart:html' as dart_html;
 import 'dart:math' as dart_math;
+import 'dart:async' as dart_async;
 
 part 'src/scheduler/scheduler.dart';