+++ lib/util.dart
@@ -7,6 +7,7 @@
 library dart_web_toolkit_util;
 
 import 'dart:html' as dart_html;
+import 'dart:async' as dart_async;
 import 'dart:uri';
 
 part 'src/util/timer.dart';